Added deleteSchedules mutation to DeleteMilestoneButton

+++ src/client/components/DeleteMilestoneButton.tsx
@@ -16,14 +16,27 @@ import { setIsBannerDisplayed } from "../features/bannerSlice.js";
 import isMostRecentStatusReportSent from "../utils/isMostRecentStatusReportSent.js";
 import getFirstCheckInDayDate from "../utils/getFirstCheckInDayDate.js";
 import { User } from "@prisma/client";
+import getHabitScheduleIds from "../utils/getHabitScheduleIds.js";
 
 export interface DeleteMilestoneButtonProps{
     milestone: MilestoneWithDetails
 }
 
 const DeleteMilestoneButton = ({milestone}: DeleteMilestoneButtonProps) => {
-    const [deleteMilestone, {isLoading: isDeleteMilestoneLoading}] = useDeleteMilestoneMutation();
-    const [deleteSchedules, {isLoading: isDeleteSchedulesLoading}] = useDeleteSchedulesMutation();
+    const [
+        deleteMilestone, 
+        {
+            isLoading: isDeleteMilestoneLoading, 
+            error: deleteMilestoneError
+        }
+    ] = useDeleteMilestoneMutation();
+    const [
+        deleteSchedules, 
+        {
+            isLoading: isDeleteSchedulesLoading,
+            error: deleteSchedulesError
+        }
+    ] = useDeleteSchedulesMutation();
 
     const toast = useToast();
     const dispatch = useDispatch();
@@ -33,17 +46,34 @@ const DeleteMilestoneButton = ({milestone}: DeleteMilestoneButtonProps) => {
     
     if (currentUser) {
         const currentUserId = currentUser.id
-        const { data, isLoading: isMilestonesLoading, error } = useGetMilestonesByUserQuery(currentUserId); 
+        const { 
+            data, 
+            isLoading: isMilestonesLoading, 
+            error 
+        } = useGetMilestonesByUserQuery(currentUserId); 
 
         const handleDeleteMilestone = async () => {
-            if (typeof error === "undefined") {
+            if (
+                typeof error === "undefined" &&
+                typeof deleteMilestoneError === "undefined" &&
+                typeof deleteSchedulesError === "undefined"
+                ) {
                 try {
+                    const scheduleIds = getHabitScheduleIds(milestone);
+
+                    const deleteSchedulesResponse = await deleteSchedules({
+                        scheduleIds
+                    }).unwrap()
+
                     const { milestone: deletedMilestone } = await deleteMilestone({
                         ownerId: currentUser.id,
                         milestoneId: milestone.id
                     }).unwrap();
     
-                    if (deletedMilestone) {
+                    if (
+                        deletedMilestone && 
+                        deleteSchedulesResponse.message === "successfully deleted"
+                        ) {
                         toast({
                             title: 'Goal Deleted',
                             description: `"${milestone.name}" has been successfully deleted`,
@@ -91,8 +121,6 @@ const DeleteMilestoneButton = ({milestone}: DeleteMilestoneButtonProps) => {
                     isClosable: true
                 })
             }
-    
-            
         }
 
         return (
